requirements.txt
streamlit
sentence-transformers
faiss-cpu
numpy
python-docx
pdf2image
pytesseract
easyocr
torch
docx
supabase>=2.0.0
python-dotenv
